fix(publish): move summarizePublishError to lib/ so next build accepts route.ts
Next.js route modules may only export HTTP handlers and a fixed set of config
fields; an extra named export (summarizePublishError) fails `next build` with
"is not a valid Route export field". Move it to lib/publish-error.ts and import
it in the route + test. No behaviour change.
